+++ game/src/components/war_game_component.tsx
@@ -7,6 +7,7 @@ import themeDefaultLight from '../style/themes/default_light';
 import HUDAlert from './alert/hud_alert';
 import WarCanvas from './canvas';
 import HUDConfirmation from './confirmation/hud_confirmation';
+import LegalDisclaimerScreen from './disclaimer/legal_disclaimer_screen';
 import FailedToStartEngineScreen from './error/failed_to_start_engine/failed_to_start_engine_screen';
 import HUDPerformance from './game/hud/debug/hud_performance';
 import EscMenu from './game/hud/esc_menu/esc_menu';
@@ -36,6 +37,7 @@ const WarGameComponent: React.FC = () => {
     return (
         <ThemeProvider theme={theme}>
             <CssBaseline enableColorScheme />
+            <LegalDisclaimerScreen />
             <FailedToStartEngineScreen />
             <LoadingScreen />
             <MenuRouter />