+++ game/src/i18next.ts
@@ -1,15 +1,16 @@
 import i18next from 'i18next';
 import { initReactI18next } from 'react-i18next';
-import countries_enUS from './locales/en-US/countries.json';
-import loading_enUS from './locales/en-US/loading.json';
-import engine_enUS from './locales/en-US/engine.json';
-import config_enUS from './locales/en-US/config.json';
-import server_list_enUS from './locales/en-US/server_list.json';
 import common_enUS from './locales/en-US/common.json';
+import config_enUS from './locales/en-US/config.json';
+import countries_enUS from './locales/en-US/countries.json';
+import credits_enUS from './locales/en-US/credits.json';
+import disclaimer_enUS from './locales/en-US/disclaimer.json';
+import engine_enUS from './locales/en-US/engine.json';
+import ingame_enUS from './locales/en-US/ingame.json';
+import loading_enUS from './locales/en-US/loading.json';
 import lobby_enUS from './locales/en-US/lobby.json';
 import parties_enUS from './locales/en-US/parties.json';
-import ingame_enUS from './locales/en-US/ingame.json';
-import credits_enUS from './locales/en-US/credits.json';
+import server_list_enUS from './locales/en-US/server_list.json';
 
 export const defaultNS = 'ns1';
 
@@ -25,6 +26,7 @@ export const resources = {
         parties: parties_enUS,
         ingame: ingame_enUS,
         credits: credits_enUS,
+        disclaimer: disclaimer_enUS,
     },
 };