+++ server/eslint.config.mjs
@@ -12,6 +12,7 @@ export default [
         rules: {
             '@typescript-eslint/no-explicit-any': ['off'],
             '@typescript-eslint/no-require-imports': ['off'],
+            '@typescript-eslint/no-unused-vars': ['warn', { argsIgnorePattern: '^_', varsIgnorePattern: '^_' }],
             'no-control-regex': ['off'],
         },
     },