+++ server/src/game/party/not_set.ts
@@ -5,7 +5,7 @@ export class PartyNotSet extends Party {
     get startingTerritories(): TerritoryCode[] | 'any' {
         throw new Error('Method not implemented.');
     }
-    calculateAllowedTurnActionsForPhase(phase: TurnPhase): TurnAllowedActions {
+    calculateAllowedTurnActionsForPhase(_phase: TurnPhase): TurnAllowedActions {
         throw new Error('Method not implemented.');
     }
     checkWinCondition(): boolean {