+++ server/src/socket/packet/lobby/game_start_cancelled.ts
@@ -1,8 +1,7 @@
-import { ServerPacket } from "../server_packet";
+import { ServerPacket } from '../server_packet';
 
-export class ServerPacketGameStartCancelled extends ServerPacket<"lGameStartCancelled"> {
-
+export class ServerPacketGameStartCancelled extends ServerPacket<'lGameStartCancelled'> {
     constructor() {
-        super("lGameStartCancelled");
+        super('lGameStartCancelled');
     }
 }