+++ server/src/socket/packet/lobby/joined_lobby.ts
@@ -1,9 +1,8 @@
-import { Lobby } from "../../../game/lobby/lobby";
-import { ServerPacket } from "../server_packet";
+import { Lobby } from '../../../game/lobby/lobby';
+import { ServerPacket } from '../server_packet';
 
-export class ServerPacketJoinedLobby extends ServerPacket<"joinedLobby"> {
-    
+export class ServerPacketJoinedLobby extends ServerPacket<'joinedLobby'> {
     constructor(lobby: Lobby) {
-        super("joinedLobby", lobby.asProtocolLobbyState);
+        super('joinedLobby', lobby.asProtocolLobbyState);
     }
 }