+++ server/src/socket/packet/lobby/left_lobby.ts
@@ -1,8 +1,7 @@
-import { ServerPacket } from "../server_packet";
+import { ServerPacket } from '../server_packet';
 
-export class ServerPacketLeftLobby extends ServerPacket<"leftLobby"> {
-    
+export class ServerPacketLeftLobby extends ServerPacket<'leftLobby'> {
     constructor(kicked?: boolean) {
-        super("leftLobby", kicked);
+        super('leftLobby', kicked);
     }
 }